pyproject.toml
[project]
name = "20a202600108-letrunganhquoc-day7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.31.0",
    "pytest>=7.0",
    "python-dotenv>=1.0",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.0",
]
